Eigenlayer (#91)
* Add blocklist factory, update exit queue

* Fix exiting assets constructor arg

* Move genesis vault legacy rewards update to _processTotalAssetsDelta

* Remove ejectUser functions, change genesis withdrawal creds to pool escrow

* Update snapshots

* Fix linting

* Move queued shares in interface

* Emit Transfer event on update exit queue

* Disable slither shadow state variable warning

* Rename ExitQueueEntered to V2ExitQueueEntered

* Revert changing withdrawal creds to pool escrow for genesis vault

* Deploy new factories, implementations to holesky

* Fix genesis vault snapshot

* Fix genesis vault fork tests

* Fix incorrect calculateExitedAssets interface outputs

* Extract Vault deposit data management to separate contract

* Remove redundant error

* Gas optimiziation

* Upgrade packages, fix coverage

* Move DepositDataManager from misc to validators

* Rename keysManager to validatorsManager

* Add version check to DepositDataManager

* Fix deprecated events

* Add multicall to the DepositDataManager

* Update snapshots

* Update IDepositDataManager interface

* Rename DepositDataManager to DepositDataRegistry

* Make _withdrawalCredentials function private

* Add state update to DepositDataRegistry

* Implement gnosis v3 vaults

* Remove redundant config

* Skip forks for gnosis tests

* Fix invalid factory addresses

* Add deposit data manager to the Gnosis vaults

* Fix gas snapshots

* Update genesis snapshots

* Update abis, snapshots

* Remove xdaiManager, rename DepositDataManager

* Fix vault assets

* Fix PriceFeeds, add stale price check

* Implement gnosis v3 vaults

* Add deposit data manager to the Gnosis vaults

* Add deposit data manager to the Gnosis vaults

* Implement eigenlayer vaults

* Extract vault EigenPod to separate contract

* Fix snapshots

* Refactor native restaking vaults

* Fix tests

* Simplify calls to eigen pods, add tests

* Remove redeemFromLtv and redeemToLtv percents

* Move exit position delay check to claimExitedAssets

* Add EthRestakeVaultFactory

* Fix tests and snapshots

* Add validators manager signature

* Audit fixes

* Fix mock contract

* Fix tests, update snapshots

* Fix snapshot

* Fix snapshots for manager signature

* Fix snapshots and test

* Remove validators manager signatures snapshots

* Add upgrade script

* Fix _computeVaultValidatorsDomain docstring

* Fix compiler version, docstring

* Validators checker (#90)

* Move exit delay check to claimExitedAssets

* Add EthValidatorsChecker

* Add tests

* Add test 'fails for vault not collateralized not deposited'

* tests 2

* fix success test

* Rework to eip-712

* add tests for check-deposit-data-root

* Cleanup

* Fix deposit registry check

* Add abi

* Refactor tests: add validatorsManagerSetup fixture

* Refactor tests: del vaultWithDepositDataManager

* Refactor tests: add getMultiProofArgs fixture

* Upd snapshots

* Add more comments

* Fix abi

* Fix tests fork

* Del loadFixture

* Add RUST_BACKTRACE: full

* Revert "Add RUST_BACKTRACE: full"

This reverts commit 01889177102e01c86e96c5990fcaad0234e709a3.

* Del EthValidatorsChecker.spec.ts

* Del ethValidatorsChecker from ethVaultFixture

* Revert "Del ethValidatorsChecker from ethVaultFixture"

This reverts commit 6ca4b465712323ecde4b589af89d13a83ecfa8f8.

* Revert "Del EthValidatorsChecker.spec.ts"

This reverts commit d43d9a2126c2dc7fc7c36a3eb96197432ccb239c.

* Add test for InvalidProof

* Log MAINNET_FORK_RPC_URL

* Revert "Log MAINNET_FORK_RPC_URL"

This reverts commit 56e657af5766467eed9b1fc490107488f38cd1f7.

* Review fixes

* Review fixes 2

* Added GnoValidatorsChecker

* Add tests for Gnosis

* Upd snapshots

* Rename abi

* Review fixes 3

* Upd snapshots

* Fix imports

* Add .python-version to gitignore

* Add deploy

* Optimize slither

* Review fixes 4

* Add validators checker address to chiado.json, holesky.json

* Replace signature with vault validators signature

* Remove createValidatorPublicKeys

* Deploy to holesky

* Deploy to chiado

* Remove unused import

* Fix EigenPodOwner verification

* Fix docstring

---------


---------

+++ contracts/interfaces/IVaultOsToken.sol
@@ -106,11 +106,11 @@ interface IVaultOsToken is IVaultState, IVaultEnterExit {
    * @param osTokenShares The number of shares to burn
    * @return assets The number of assets burned
    */
-  function burnOsToken(uint128 osTokenShares) external returns (uint256 assets);
+  function burnOsToken(uint256 osTokenShares) external returns (uint256 assets);
 
   /**
    * @notice Liquidates a user position and returns the number of received assets.
-   *         Can only be called when health factor is below 1.
+   *         Can only be called when health factor is below 1 by the liquidator.
    * @param osTokenShares The number of shares to cover
    * @param owner The address of the position owner to liquidate
    * @param receiver The address of the receiver of the liquidated assets
@@ -118,7 +118,7 @@ interface IVaultOsToken is IVaultState, IVaultEnterExit {
   function liquidateOsToken(uint256 osTokenShares, address owner, address receiver) external;
 
   /**
-   * @notice Redeems osToken shares for assets. Can only be called when health factor is above redeemFromHealthFactor.
+   * @notice Redeems osToken shares for assets. Can only be called when health factor is above redeemFromHealthFactor by the redeemer.
    * @param osTokenShares The number of osToken shares to redeem
    * @param owner The address of the position owner to redeem from
    * @param receiver The address of the receiver of the redeemed assets
